use i64 for storage

+++ SCHEMA.sql
@@ -1,15 +1,15 @@
 CREATE TABLE IF NOT EXISTS [Clients] (
     id          INTEGER PRIMARY KEY,
-    available   REAL NOT NULL,
-    held        REAL NOT NULL,
+    available   BIGINT NOT NULL,
+    held        BIGINT NOT NULL,
     locked      BOOLEAN NOT NULL
 );
 
 CREATE TABLE IF NOT EXISTS [Transactions] (
-    id                  INTEGER PRIMARY KEY,
-    [type]              TEXT NOT NULL,
+    id                      INTEGER PRIMARY KEY,
+    [type]                  TEXT NOT NULL,
     client_id              INTEGER NOT NULL,
-    amount              REAL,
+    amount                  BIGINT,
 	FOREIGN KEY(client_id) REFERENCES Clients(id)
 );
